phase2/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0
python-dotenv==1.0.0
azure-ai-contentsafety==1.0.0
azure-ai-textanalytics==5.3.0
azure-identity==1.14.0
azure-storage-blob==12.18.2
openai==1.1.1
pydantic==2.3.0
langchain==0.0.312
langchain-community==0.0.6
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
loguru==0.7.2
python-multipart==0.0.6 